_webservice/upd/update2.json: 12.2.53 AxTools_x64/Forms/MainForm.cs: <Notify.Balloon> --> <Notify.TrayPopup> AxTools_x64/Forms/WowRadar.cs: added ability to draw local player's rotation and 40m circle after drawing of objects, players and NPC AxTools_x64/Forms/WowRadarOptions.cs: refactoring + added <checkBoxPlayerArrowOnTop> (option for ability to draw local player's rotation and 40m circle after drawing of objects, players and NPC) AxTools_x64/Helpers/MemoryManagement/Imports.cs: <VirtualAllocEx>: second parameter is <IntPtr> now AxTools_x64/Helpers/MemoryManagement/MemoryManager.cs: refactoring AxTools_x64/Helpers/Notify.cs: <Balloon> is deleted AxTools_x64/Helpers/Settings.cs: new field: <WoWRadarShowLocalPlayerRotationArrowOnTop> AxTools_x64/Services/AddonsBackup.cs: refactoring + <Notify.Balloon> --> <Notify.TrayPopup> AxTools_x64/WoW/PluginSystem/API/GameFunctions.cs: refactoring + <Notify.Balloon> --> <Notify.TrayPopup> AxTools_x64/WoW/PluginSystem/API/Utilities.cs: new method: <LoadJSON> AxTools_x64/WoW/PluginSystem/PluginManager.cs: added ability to check dependencies of plugins: <CheckDependencies> AxTools_x64/WoW/PluginSystem/Plugins/FlagReturner.cs: debug code is deleted AxTools_x64/WoW/WowBuildInfoX64.cs: WoW build 24330 (except UIFrame) WoWPlugins/WoWPlugin_PathPlayer/PathPlayer.cs: this addon now is just a wrapper for <LibNavigator> WoWPlugins/WoWPlugin_PathPlayer/Settings.cs: <LoopPath> and <StartFromNearestPoint> are deleted WoWPlugins/WoWPlugin_PathPlayer/SettingsForm.cs: <checkBoxLoopPath> and <checkBoxStartFromNearestPoint> are deleted AxTools_x64/Updater/UpdateInfo2.cs: just renaming file: <UpdateInfo.cs> --> <UpdateInfo2.cs> AxTools_x64/WoW/PluginSystem/IPlugin2.cs: derived from <IPlugin>. One field: string[] Dependencies WoWPlugins/LibNavigator/DoAction.cs: initial commit, just copy of <PathPlayer> WoWPlugins/LibNavigator/DoActionType.cs: copy of <PathPlayer> + added new fields: <SetPrecision3D>, <SetLoopPath> and <SetStartFromNearestPoint> WoWPlugins/LibNavigator/LibNavigator.cs: copy of <PathPlayer> AxTools_x64/Updater/UpdateInfo.cs: just renaming file: <UpdateInfo.cs> --> <UpdateInfo2.cs> WoWPlugins/WoWPlugin_PathPlayer/DoAction.cs: functionality is moved to <LibNavigator> WoWPlugins/WoWPlugin_PathPlayer/DoActionType.cs: functionality is moved to <LibNavigator> WoWPlugins/LibNavigator/LibNavigator.csproj: new addon WoWPlugins/LibNavigator/Properties/AssemblyInfo.cs: new addon

+++ AxTools_x64/Helpers/MemoryManagement/Imports.cs
@@ -131,7 +131,7 @@ namespace AxTools.Helpers.MemoryManagement
         /// </returns>
         /// <remarks>Created 2012-02-15</remarks>
         [DllImport("kernel32", EntryPoint = "VirtualAllocEx")]
-        internal static extern IntPtr VirtualAllocEx(SafeMemoryHandle hProcess, uint dwAddress, int nSize, MemoryAllocationType dwAllocationType, MemoryProtectionType dwProtect);
+        internal static extern IntPtr VirtualAllocEx(SafeMemoryHandle hProcess, IntPtr dwAddress, int nSize, MemoryAllocationType dwAllocationType, MemoryProtectionType dwProtect);
 
         /// <summary>
         ///     Releases, decommits, or releases and decommits a region of memory within the virtual address space of a specified
